fix (runtime): arraysDiffSequence
